**核心摘要**
*   \[Refactor]\[RunScriptExtension] 优化颜色匹配和坐标计算逻辑

**所有的非隐藏文件级变更说明**

*-RunScriptExtension.kt*

*   重构颜色匹配逻辑，优化颜色匹配失败时的日志输出
*   优化坐标计算逻辑，允许坐标计算误差
*   修复labelPos大于0时，坐标计算错误的问题

*-ScriptSetInfoDao.kt*

*   新增根据setId更新setValue的数据库操作

*-autodaily.cpp*

*   修改颜色数据类型为int
*   更新ocrResMethod的方法签名

*-common.h*

*   修改颜色数据类型为int

*-CommandImpl.kt*

*   新增FinishFlowId指令，用于结束指定流程ID
*   新增AddPosById指令，用于增加指定Action的labelPos
*   新增DropdownMenuNext指令，用于循环切换下拉菜单选项
*   新增MinusPosById指令，用于减少指定Action的labelPos
*   新增Reboot指令，用于重启应用
*   新增RmSkipAcIdList指令，用于移除跳过的Action ID列表

*-CrnnNet.h*

*   修改colorMapping返回值类型为int

*-LogScreen.kt*

*   优化日志显示，使用bodySmall样式
*   修复日志未开启时，显示错误提示的问题

*-AndroidManifest.xml*

*   添加硬件加速属性

*-RetrofitCreate.kt*

*   修改Retrofit的BaseUrl

*-ActionString.kt*

*   新增POS\_MINUS、DROP\_SET\_NEXT、REBOOT、RM\_SKIP\_ACIDS常量

*-ScriptActionInfo.kt*

*   修改hsv属性的数据类型为Set<Int>

*-ForegroundService.kt*

*   降低通知优先级

*-RunScript.kt*

*   优化初始化全局设置的逻辑
*   修复当前时间段已执行，跳过：{si.scriptName} 的bug
*   优化tryAction方法，增加对FinishFlowId、AddPosById、DropdownMenuNext、MinusPosById、Reboot等指令的处理
*   修改setScriptSetStatus方法，增加flowId参数
*   优化action初始化函数，增加对POS\_MINUS、DROP\_SET\_NEXT、REBOOT、RM\_SKIP\_ACIDS指令的处理

*-Lom.kt*

*   优化日志写入，增加文件大小限制

*-RadioButtonSettingItem.kt*

*   修改布局

*-OcrResult.kt*

*   修改colorSet属性的数据类型为Set<Int>
*   修改colorArr属性的数据类型为IntArray

*-HomeViewModel.kt*

*   修复删除脚本时，未删除对应Action的问题

*-CrnnNet.cpp*

*   优化颜色统计和映射逻辑

+++ app/src/main/cpp/common.h
@@ -14,7 +14,7 @@ struct TextLine {
     std::vector<short> label;
     std::string text;
     std::vector<float> charScores;
-    std::vector<short> color;
+    std::vector<int> color;
 };
 struct Object
 {
